USB HOST, MSC: добавил функции чтния/записи для диской большизЮ чем 2 ТБ

+++ lib/Middlewares/ST/STM32_USB_Host_Library/Class/MSC/Inc/usbh_msc.h
@@ -181,10 +181,10 @@ USBH_StatusTypeDef USBH_MSC_GetLUNInfo(USBH_HandleTypeDef *phost, uint8_t lun,
                                        MSC_LUNTypeDef *info);
 
 USBH_StatusTypeDef USBH_MSC_Read(USBH_HandleTypeDef *phost, uint8_t lun,
-                                 uint32_t address, uint8_t *pbuf, uint32_t length);
+                                 uint64_t address, uint8_t *pbuf, uint32_t length);
 
 USBH_StatusTypeDef USBH_MSC_Write(USBH_HandleTypeDef *phost, uint8_t lun,
-                                  uint32_t address, uint8_t *pbuf, uint32_t length);
+                                  uint64_t address, uint8_t *pbuf, uint32_t length);
 /**
   * @}
   */
